All Products: Use block.json to register
